+++ Nos/Views/Profile/ProfileFeedType.swift
@@ -1,4 +1,5 @@
 import CoreData
+import SwiftUI
 
 /// An enumeration of the different feed algorithms the user can choose to view on the Profile screen.
 enum ProfileFeedType {
@@ -19,5 +20,34 @@ enum ProfileFeedType {
         }
         
         return Filter(authorKeys: [author.hexadecimalPublicKey ?? "error"], kinds: kinds)
+    }
+}
+
+extension ProfileFeedType: NosSegmentedPickerItem {
+    var id: String {
+        switch self {
+        case .activity:
+            "activity"
+        case .notes:
+            "notes"
+        }
+    }
+    
+    var titleKey: LocalizedStringKey {
+        switch self {
+        case .activity:
+            "activity"
+        case .notes:
+            "notes"
+        }
+    }
+    
+    var image: Image {
+        switch self {
+        case .activity:
+            Image.profileFeed
+        case .notes:
+            Image.profilePosts
+        }
     }
 }